# public/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Layanan KP",
        "description": "Dokumentasi API Validasi Logbook Mahasiswa",
        "contact": {
            "email": "admin@unand.ac.id"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/pembimbing": {
            "post": {
                "tags": [
                    "Pembimbingan"
                ],
                "summary": "Menentukan dosen pembimbing baru.",
                "description": "Mencatat penunjukan dosen pembimbing beserta judul tugas untuk mahasiswa.",
                "operationId": "tentukanPembimbing",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "student_id",
                                    "lecturer_id",
                                    "judul"
                                ],
                                "properties": {
                                    "student_id": {
                                        "description": "ID Mahasiswa yang akan dibimbing.",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "lecturer_id": {
                                        "description": "ID Dosen pembimbing.",
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "judul": {
                                        "description": "Judul tugas yang akan dibimbing.",
                                        "type": "string",
                                        "example": "Rancangan Sistem Informasi Akademik"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dosen pembimbing berhasil ditentukan",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Dosen pembimbing berhasil ditentukan"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/surat-tugas": {
            "post": {
                "tags": [
                    "Surat Tugas"
                ],
                "summary": "Membuat dan menyimpan surat tugas pembimbing.",
                "description": "Membuat file surat tugas (simulasi .txt) dan mencatatnya ke database.",
                "operationId": "buatSuratTugas",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "supervision_id"
                                ],
                                "properties": {
                                    "supervision_id": {
                                        "description": "ID data Supervision yang akan dibuatkan surat tugas.",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Surat tugas pembimbing berhasil dibuat"
                    },
                    "404": {
                        "description": "Data pembimbing tidak ditemukan"
                    }
                }
            }
        },
        "/surat-tugas/{id}": {
            "get": {
                "tags": [
                    "Surat Tugas"
                ],
                "summary": "Mengunduh file surat tugas.",
                "description": "Mengunduh file surat tugas yang tersimpan di storage lokal berdasarkan ID Surat Tugas.",
                "operationId": "downloadSuratTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID SuratTugas yang akan diunduh.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengunduh file.",
                        "content": {
                            "application/octet-stream": {}
                        }
                    },
                    "404": {
                        "description": "File surat tugas tidak ditemukan"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Surat Tugas"
                ],
                "summary": "Menghapus surat tugas dan file terkait.",
                "description": "Menghapus record SuratTugas dari database dan file dari storage.",
                "operationId": "deleteSuratTugas",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID SuratTugas yang akan dihapus.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Surat tugas pembimbing berhasil dihapus"
                    },
                    "404": {
                        "description": "Surat tugas pembimbing tidak ditemukan"
                    }
                }
            }
        },
...
        "/pembimbing/{id}": {
            "put": {
                "tags": [
                    "Pembimbingan"
                ],
                "summary": "Memperbarui dosen pembimbing dan/atau judul tugas.",
                "description": "Memperbarui data pembimbing (Supervision), menghapus surat tugas lama, dan membuat yang baru secara otomatis.",
                "operationId": "updatePembimbing",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID data Supervision yang akan diperbarui.",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lecturer_id"
                                ],
                                "properties": {
                                    "lecturer_id": {
                                        "description": "ID Dosen pembimbing yang baru.",
                                        "type": "integer",
                                        "example": 15
                                    },
                                    "judul": {
                                        "description": "Judul tugas yang diperbarui (optional).",
                                        "type": "string",
                                        "example": "Perancangan Ulang Sistem Informasi"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dosen pembimbing berhasil diperbarui dan surat tugas baru dibuat"
                    },
                    "404": {
                        "description": "Data pembimbing tidak ditemukan"
                    }
                }
            }
        },
...
    },
...
}